re-import classes to porting lib

+++ src/main/java/committee/nova/mods/avaritia/api/utils/math/RayTracer.java
@@ -14,7 +14,7 @@ import net.minecraft.world.phys.Vec3;
 public class RayTracer {
 
     public static Vec3 getCorrectedHeadVec(Player player) {
-        return player.position().add(0.0D, (double) player.getEyeHeight(), 0.0D);
+        return player.position().add(0.0D, player.getEyeHeight(), 0.0D);
     }
 
     public static Vec3 getStartVec(Player player) {
